corrected functiondeclaration, strncmp"

+++ ft_memcpy.c
@@ -10,11 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stddef.h>
+#include "libft.h"
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int			i;
+	size_t		i;
 	char		*cdest;
 	const char	*csrc;
