+++ ft_memmove.c
@@ -10,11 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stddef.h>
+#include "libft.h"
 
 void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	int			i;
+	size_t		i;
 	char		*c_dest;
 	const char	*c_src;
